(BSR): remove useless component
